Update to re-use build result type from build-utils
